requirements.txt
django = 4.1.7
psycopg2 = 2.9.5
djangorestframework = 3.14.0
django-cors-headers = 3.14.0